requirements.txt
torch
torchvision
transformers
datasets
scikit-learn
accelerate
pillow
numpy